% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_geom_distance_at_de}
\alias{add_geom_distance_at_de}
\title{Add geom distance at De}
\usage{
add_geom_distance_at_de(gp, distance_at_de)
}
\arguments{
\item{gp}{A ggplot object returned from \code{dr_plot}.}

\item{distance_at_de}{A tibble of the output of \code{dr_distance_at_de}.}
}
\value{
A ggplot2 object.
}
\description{
Add layer of predicted distance at target DE to ggplot created by \code{dr_plot()}
}
\examples{
\dontrun{
dr_plot(range_test) \%>\%
add_geom_predicted(predicted) \%>\%
add_geom_distance_at_de(distance_at_de)
}
}
\seealso{
Other plot: 
\code{\link{add_geom_predicted}()},
\code{\link{dr_plot}()}
}
\concept{plot}
