% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_geom_predicted}
\alias{add_geom_predicted}
\title{Add geom predicted}
\usage{
add_geom_predicted(gp, predicted)
}
\arguments{
\item{gp}{A ggplot object returned from \code{dr_plot}.}

\item{predicted}{A tibble of the output of \code{dr_predict}.}
}
\value{
A ggplot2 object.
}
\description{
Add layer of predicted DE at distance to ggplot created by \code{dr_plot()}
}
\examples{
\dontrun{
dr_plot(range_test) \%>\% add_geom_predicted(predicted)
}
}
\seealso{
Other plot: 
\code{\link{add_geom_distance_at_de}()},
\code{\link{dr_plot}()}
}
\concept{plot}
