% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.drfit}
\alias{augment.drfit}
\title{Get Augmented Data from drfit Object}
\usage{
\method{augment}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the augmented data.
}
\description{
Get a tibble of the original data with augmentation.
}
\examples{
fit <- dr_fit(detrange::range_obs)
augment(fit)
}
\seealso{


Other generics: 
\code{\link{glance.drfit}()},
\code{\link{tidy.drfit}()}
}
\concept{generics}
