% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{converged.drfit}
\alias{converged.drfit}
\title{Get Convergence of drfit Object}
\usage{
\method{converged}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A flag indicating convergence.
}
\description{
Get Convergence of drfit Object
}
