% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R
\docType{data}
\name{detrange-ggproto}
\alias{detrange-ggproto}
\alias{StatDrPoint}
\alias{GeomDrPoint}
\alias{GeomDrRibbon}
\alias{GeomDrErrorbar}
\title{ggproto Classes for Plotting Detection Range}
\format{
An object of class \code{StatDrPoint} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomDrPoint} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 1.

An object of class \code{GeomDrRibbon} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{GeomDrErrorbar} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatDrPoint

GeomDrPoint

GeomDrRibbon

GeomDrErrorbar
}
\description{
ggproto Classes for Plotting Detection Range
}
\seealso{
\code{\link[ggplot2:ggproto]{ggplot2::ggproto()}}
}
\keyword{datasets}
