% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrange-package.R
\docType{package}
\name{detrange-package}
\alias{detrange}
\alias{detrange-package}
\title{detrange: Estimate Detection Range from Passive Acoustic Telemetry}
\description{
Estimate detection range from passive acoustic telemetry range test data using a Bayesian mixed effects model.
}
\author{
\strong{Maintainer}: Seb Dalgarno \email{dalgarno@unbc.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

}
\keyword{internal}
