% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analyse}
\alias{dr_analyse}
\alias{dr_analyze}
\title{Analyse model}
\usage{
dr_analyse(data, de_target = 0.5, priors = rsi_priors, nthin = 10L)
}
\arguments{
\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}

\item{de_target}{A number of the target detection efficiency (from 0 to 1) to estimate distance at.}

\item{priors}{A named list of the priors for each model term \code{bIntercept}, \code{bMidpoint}, \code{sInterceptStation}, \code{sMidpointStation}.}

\item{nthin}{A number of the number of thins.}
}
\value{
A mbr model object.
}
\description{
Analyse detection range test data with a Bayesian mixed-effects model. Each station has a random slope and intercept.
Note that with fewer data points per station, priors will have strong influence on the posterior.
The default priors are based on understanding of typical detection range.
}
\seealso{
Other analysis: 
\code{\link{dr_analysis_coef}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()},
\code{\link{dr_analysis_predict}()}
}
\concept{analysis}
