% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analysis_coef}
\alias{dr_analysis_coef}
\title{Get model coefficients}
\usage{
dr_analysis_coef(analysis, include_random = FALSE)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}

\item{include_random}{A flag indicating whether to include coefficients for random terms.}
}
\value{
A tibble of the coefficients.
}
\description{
Get coefficients from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()},
\code{\link{dr_analysis_predict}()}
}
\concept{analysis}
