% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analysis_glance}
\alias{dr_analysis_glance}
\title{Get analysis glance summary}
\usage{
dr_analysis_glance(analysis)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A tibble of the glance summary.
}
\description{
Get analysis glance summary from model object output of \code{dr_analyse}.
}
\seealso{
Other model: 
\code{\link{dr_analyse}()},
\code{\link{dr_analysis_coef}()},
\code{\link{dr_analysis_midpoint}()},
\code{\link{dr_analysis_predict}()},
\code{\link{jags_template_rsi}()}
}
\concept{model}
