% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R, R/predict.R
\name{dr_analysis_predict}
\alias{dr_analysis_predict}
\title{Predict detection range}
\usage{
dr_analysis_predict(analysis, distance_seq = NULL, by_station = TRUE)

dr_analysis_predict(analysis, distance_seq = NULL, by_station = TRUE)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}

\item{distance_seq}{A vector of the distances to make predictions for.}
}
\value{
A tibble of the coefficients.

A tibble of the coefficients.
}
\description{
Predict detection range from model object output of \code{dr_analyse}.

Predict detection range from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_analysis_coef}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()}

Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_analysis_coef}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()}
}
\concept{analysis}
