% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_coef}
\alias{dr_coef}
\title{Get model coefficients}
\usage{
dr_coef(analysis)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A tibble of the coefficients.
}
\description{
Get coefficients from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict}()},
\code{\link{dr_target}()}
}
\concept{analysis}
