% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{dr_distance_at_de}
\alias{dr_distance_at_de}
\title{Predict distance at target DE}
\usage{
dr_distance_at_de(
  analysis,
  de_target = 0.5,
  by = "Station",
  conf_level = 0.95,
  estimate = median
)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}

\item{de_target}{A number of the target detection efficiency (from 0 to 1) to estimate distance at.}

\item{by}{A string of the grouping variable to derive predictions for.}

\item{conf_level}{A number between 0 and 1 of the confidence level. Default is 0.95.}

\item{estimate}{A function to calculate the estimate with. The default is median.}
}
\value{
A tibble of the coefficients.
}
\description{
Predict distance at target detection efficiency from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_coef}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict}()}
}
\concept{analysis}
