% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_glance}
\alias{dr_glance}
\title{Get analysis glance summary}
\usage{
dr_glance(analysis)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A tibble of the glance summary.
}
\description{
Get analysis glance summary from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_coef}()},
\code{\link{dr_predict}()},
\code{\link{dr_target}()}
}
\concept{analysis}
