% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{dr_model_range}
\alias{dr_model_range}
\title{Estimate detection range.}
\usage{
dr_model_range(data)
}
\arguments{
\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}
}
\value{
The model output object.
}
\description{
Estimate detection range from range test data.
}
\details{
Returns output of a JAGS Bayesian mixed-effects model with random intercept and slope for each station.
Note that with fewer data points per station, sensible default priors will have strong influence on the posterior.
}
\examples{
\dontrun{
dr_model_range(range_test)
}
}
\concept{model}
