% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot}
\alias{dr_plot}
\alias{dr_plot.data.frame}
\alias{dr_plot.drfit}
\title{Plot Predicted Detection Range}
\usage{
dr_plot(x, ...)

\method{dr_plot}{data.frame}(
  x,
  facet = TRUE,
  xlab = "Distance",
  ylab = "Proportion of Pings Detected",
  ...
)

\method{dr_plot}{drfit}(
  x,
  facet = TRUE,
  xlab = "Distance",
  ylab = "Propotion of Pings Detected",
  ...
)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}

\item{facet}{A flag indicating whether to facet plot by Station.}

\item{xlab}{A string of the x-axis title.}

\item{ylab}{A string of the y-axis title.}
}
\description{
Generic function to plot the predicted detection range.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: method to plot predicted detection range from data.frame object

\item \code{drfit}: method to plot predicted detection range from drfit object
}}

\examples{
dr_plot(detrange::range_obs)
fit <- dr_fit(detrange::range_obs)
dr_plot(fit)
}
