% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot_ob}
\alias{dr_plot_ob}
\title{Plot predicted detection range and midpoint}
\usage{
dr_plot_ob(data)
}
\arguments{
\item{data}{A data.frame of the detection range data (see \code{\link{range_obs}})}
}
\value{
A ggplot2 object.
}
\description{
Plot predicted detection range and midpoint from analysis output of \code{dr_fit},
}
\examples{
\dontrun{
dr_plot(detrange::range_obs)
}
}
\seealso{
Other plot: 
\code{\link{add_geom_predicted_distance}()},
\code{\link{add_geom_predicted}()}
}
\concept{plot}
