% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-observed.R
\name{dr_plot_range_test}
\alias{dr_plot_range_test}
\title{Plot range test}
\usage{
dr_plot_range_test(data)
}
\arguments{
\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}
}
\value{
A ggplot2 object.
}
\description{
Plot observed range test data
}
\examples{
\dontrun{
dr_plot_range_test(detrange::range_test)
}
}
\concept{plot}
