% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{dr_predict}
\alias{dr_predict}
\title{Predict detection range}
\usage{
dr_predict(
  x,
  by = "Station",
  distance_seq = NULL,
  conf_level = 0.95,
  estimate = median
)
}
\arguments{
\item{x}{The object.}

\item{by}{A string of the grouping variable to derive predictions for.}

\item{distance_seq}{A vector of the distances to make predictions for.}

\item{conf_level}{A number between 0 and 1 of the confidence level. Default is 0.95.}

\item{estimate}{A function to calculate the estimate with. The default is median.}
}
\value{
A tibble of the coefficients.
}
\description{
Predict detection efficiency at distance sequence from model object output of \code{dr_fit}.
}
\seealso{
Other analysis: 
\code{\link{dr_coef}()},
\code{\link{dr_distance_at_de}()},
\code{\link{dr_glance}()}
}
\concept{analysis}
