% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{dr_predict_de}
\alias{dr_predict_de}
\title{Predict detection efficiency}
\usage{
dr_predict_de(
  x,
  distance,
  by = "Station",
  conf_level = 0.95,
  estimate = median
)
}
\arguments{
\item{x}{The object.}

\item{distance}{A vector of the distances to predict detection efficiency at.
If NULL, a vector of length 30 spanning the range of the Distance in the data will be generated.}

\item{by}{A string of the column indicating group to derive predictions at.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\value{
A tibble of the coefficients.
}
\description{
Predict detection efficiency at specified distance(s).
}
\seealso{
Other analysis: 
\code{\link{dr_coef}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict_distance}()}
}
\concept{analysis}
