% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{esr.drfit}
\alias{esr.drfit}
\title{Get Effective Sample Rate of drfit Object}
\usage{
\method{esr}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of chains.
}
\description{
Get Effective Sample Rate of drfit Object
}
