% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{estimates.drfit}
\alias{estimates.drfit}
\title{Estimates for drfit Object}
\usage{
\method{estimates}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A named list of the estimates.
}
\description{
Gets a named list of the estimated values by term.
}
\examples{
fit <- dr_fit(detrange::range_test)
estimates(fit)
}
\seealso{
\code{\link[=tidy.drfit]{tidy.drfit()}}
}
