% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.drfit}
\alias{glance.drfit}
\title{Get a Glance Summary of drfit Object}
\usage{
\method{glance}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the glance summary.
}
\description{
Get a tibble of a one-row summary of the model fit.
}
\examples{
fit <- dr_fit(detrange::range_obs)
glance(fit)
}
\seealso{


Other generics: 
\code{\link{augment.drfit}()},
\code{\link{tidy.drfit}()}
}
\concept{generics}
