% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{nchains.drfit}
\alias{nchains.drfit}
\title{Get Number of Chains from drfit Object}
\usage{
\method{nchains}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of chains.
}
\description{
Get Number of Chains from drfit Object
}
