% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{niters.drfit}
\alias{niters.drfit}
\title{Get Number of Iterations from drfit Object}
\usage{
\method{niters}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of iterations.
}
\description{
Get Number of Iterations from drfit Object
}
