% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{npars.drfit}
\alias{npars.drfit}
\title{Get Number of Parameters from drfit Object}
\usage{
\method{npars}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of parameters.
}
\description{
Get Number of Parameters from drfit Object
}
