% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{nterms.drfit}
\alias{nterms.drfit}
\title{Get Number of Terms from drfit Object}
\usage{
\method{nterms}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of the number of terms.
}
\description{
Get Number of Terms from drfit Object
}
