% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{pars.drfit}
\alias{pars.drfit}
\title{Get Parameters from drfit Object}
\usage{
\method{pars}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A vector of the parameter names.
}
\description{
Get Parameters from drfit Object
}
