% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.drfit}
\alias{predict.drfit}
\title{Predict Detection Efficiency}
\usage{
\method{predict}{drfit}(
  object,
  distance = NULL,
  by = "Station",
  conf_level = 0.95,
  estimate = median,
  ...
)
}
\arguments{
\item{object}{The object.}

\item{distance}{A vector of the distances to predict detection efficiency at.
If NULL, a vector of length 30 spanning the range of the Distance in the data will be generated.}

\item{by}{A string of the column indicating group to derive predictions at.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{...}{Unused parameters.}
}
\description{
A wrapper on \code{\link[=dr_predict_de]{dr_predict_de()}} that by default predicts
detection efficiency over the range of distance in data.
}
\details{
It is useful for plotting purposes.
}
\examples{
fit <- dr_fit(detrange::range_obs)
predict(fit)
}
\seealso{
\code{\link[=dr_predict_de]{dr_predict_de()}}
}
