% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_obs}
\alias{range_obs}
\title{Simulated range test data}
\format{
A data.frame.
}
\usage{
range_obs
}
\description{
Simulated range test data
}
\examples{
range_obs
}
\seealso{
Other data: 
\code{\link{range_pred}}
}
\concept{data}
\keyword{datasets}
