% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_pred}
\alias{range_pred}
\title{Simulated detection range prediction}
\format{
A data.frame.
}
\usage{
range_pred
}
\description{
Simulated detection range prediction
}
\examples{
range_pred
}
\seealso{
Other data: 
\code{\link{range_obs}}
}
\concept{data}
\keyword{datasets}
