% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_test}
\alias{range_test}
\title{Simulated range test data}
\format{
A data.frame.
}
\usage{
range_test
}
\description{
Simulated range test data
}
\examples{
range_test
}
\concept{data}
\keyword{datasets}
