% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/coef.R, R/glance.R, R/predict.R,
%   R/summary.R, R/tidy.R, R/universals.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{coef}
\alias{glance}
\alias{predict}
\alias{nobs}
\alias{tidy}
\alias{estimates}
\alias{nchains}
\alias{npars}
\alias{nterms}
\alias{niters}
\alias{rhat}
\alias{esr}
\alias{converged}
\alias{pars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats]{nobs}}, \code{\link[stats]{predict}}}

  \item{universals}{\code{\link[universals]{converged}}, \code{\link[universals]{esr}}, \code{\link[universals]{estimates}}, \code{\link[universals]{nchains}}, \code{\link[universals]{niters}}, \code{\link[universals]{npars}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pars}}, \code{\link[universals]{rhat}}}
}}

