% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universals.R
\name{rhat.drfit}
\alias{rhat.drfit}
\title{Get Rhat of drfit Object}
\usage{
\method{rhat}{drfit}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused parameters.}
}
\value{
A number of rhat value.
}
\description{
Get Rhat of drfit Object
}
