% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.drfit}
\alias{tidy.drfit}
\title{Get Tidy Tibble from drfit Object.}
\usage{
\method{tidy}{drfit}(x, conf_level = 0.95, estimate = median, random_effects = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{random_effects}{A flag indicating whether to include random effect terms.}

\item{...}{Unused parameters.}
}
\value{
A tibble of the tidy coefficient summary.
}
\description{
Get a tidy tibble of the coefficient estimates and
confidence intervals from model fit.
}
\examples{
fit <- dr_fit(detrange::range_test)
tidy(fit)
}
\seealso{
\code{\link[=coef.drfit]{coef.drfit()}}

Other generics: 
\code{\link{augment.drfit}()},
\code{\link{glance.drfit}()}
}
\concept{generics}
