% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R
\docType{data}
\name{detrange-ggproto}
\alias{detrange-ggproto}
\alias{StatObspoint}
\alias{GeomObspoint}
\title{ggproto Classes for Plotting Detection Range}
\format{
An object of class \code{StatObspoint} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomObspoint} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 1.
}
\usage{
StatObspoint

GeomObspoint
}
\description{
ggproto Classes for Plotting Detection Range
}
\seealso{
\code{\link[ggplot2:ggproto]{ggplot2::ggproto()}}
}
\keyword{datasets}
