% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analyse}
\alias{dr_analyse}
\alias{dr_analyze}
\title{Analyse model}
\usage{
dr_analyse(data, priors = rsi_priors, nthin = 10L)
}
\arguments{
\item{data}{A tibble of the range test data.}

\item{priors}{A list of the priors for each model term.}

\item{nthin}{A number of the number of thins.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}
}
\value{
A mbr model object.
}
\description{
Analyse detection range test data with a Bayesian mixed-effects model. Each station has a random slope and intercept.
Note that with fewer data points per station, priors will have strong influence on the posterior.
The default priors are based on understanding of typical detection range.
}
\seealso{
Other model: 
\code{\link{dr_analysis_coef}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()},
\code{\link{dr_analysis_predict}()},
\code{\link{jags_template_rsi}()}
}
\concept{model}
