% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analyse}
\alias{dr_analyse}
\alias{dr_analyze}
\title{Analyse model}
\usage{
dr_analyse(data, nthin = 10L, priors = NULL, quiet = TRUE)
}
\arguments{
\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}

\item{nthin}{A number of the number of thins.}

\item{priors}{A named list of the priors for each model term \code{bIntercept}, \code{bMidpoint}, \code{sInterceptStation}, \code{sMidpointStation}.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}
}
\value{
A list of the jags model object and mcmcr samples.
}
\description{
Analyse detection range test data with a Bayesian model fit with JAGS.
Each station has a random slope and intercept.
}
\seealso{
Other analysis: 
\code{\link{dr_coef}()},
\code{\link{dr_distance_at_de}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict}()}
}
\concept{analysis}
