% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_analysis_coef}
\alias{dr_analysis_coef}
\alias{dr_analyze}
\title{Get model coefficients}
\usage{
dr_analysis_coef(analysis, include_random = TRUE)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A tibble of the coefficients.
}
\description{
Get coefficients from model object output of \code{dr_analyse}.
}
\seealso{
Other model: 
\code{\link{dr_analyse}()},
\code{\link{dr_analysis_glance}()},
\code{\link{dr_analysis_midpoint}()},
\code{\link{dr_analysis_predict}()},
\code{\link{jags_template_rsi}()}
}
\concept{model}
