% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_coef}
\alias{dr_coef}
\title{Get model coefficients}
\usage{
dr_coef(analysis, conf_level = 0.95, estimate = median, random_effects = FALSE)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}

\item{conf_level}{A number between 0 and 1 of the confidence level. Default is 0.95.}

\item{estimate}{A function to calculate the estimate with. The default is median.}

\item{random_effects}{A flag indicating whether to include random effects.}
}
\value{
A tibble of the coefficients.
}
\description{
Get coefficients from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_distance_at_de}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict}()}
}
\concept{analysis}
