% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dr_coef}
\alias{dr_coef}
\title{Get model coefficients}
\usage{
dr_coef(x, conf_level = 0.95, estimate = median, random_effects = FALSE)
}
\arguments{
\item{x}{The object.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{random_effects}{A flag indicating whether to include random effect terms.}
}
\value{
A tibble of the coefficients.
}
\description{
Get coefficients from model object output of \code{dr_fit}.
}
\seealso{
Other analysis: 
\code{\link{dr_glance}()},
\code{\link{dr_predict_de}()},
\code{\link{dr_predict_distance}()}
}
\concept{analysis}
