% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dr_fit}
\alias{dr_fit}
\title{Fit Detection Range Model}
\usage{
dr_fit(
  data,
  nthin = 10,
  min_random_slope = 5,
  min_random_intercept = 5,
  priors = NULL,
  quiet = TRUE,
  seed = .rndm_seed()
)
}
\arguments{
\item{data}{A data.frame of the detection range data (see \code{\link{range_obs}})}

\item{nthin}{A whole number of the number of thins.}

\item{min_random_slope}{A whole number of the minimum number of Stations
required to fit slope parameter as a random effect.}

\item{min_random_intercept}{A whole number of the minimum number of Stations
required to fit intercept parameter as a random effect.}

\item{priors}{A named list of the parameter names and priors (in BUGS language) including some subset of the parameters required by the model.
Any parameters not included will fall back to defaults in \code{priors()}. If NULL, all default priors are used.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}

\item{seed}{A positive whole number of the seed.}
}
\value{
A list of the jags model object and mcmcr samples.
}
\description{
Fit detection range model using JAGS.
If the number of stations is < min_random, a fixed-effects model is fit.
Otherwise, a mixed-effects model is fit with random slope for each station.
If random_intercept = TRUE, a random intercept is also fit for each station.
}
\concept{model}
