% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dr_fit}
\alias{dr_fit}
\title{Analyse model}
\usage{
dr_fit(
  data,
  nthin = 10,
  min_random = 5,
  random_intercept = FALSE,
  priors = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data}{A tibble of the detection range data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}

\item{nthin}{A whole number of the number of thins.}

\item{min_random}{A whole number of the minimum number of Stations required to fit a mixed-effects model.}

\item{random_intercept}{A flag indicating whether to model intercept as a random effect.
In the case of a fixed-effects model, this will be ignored.}

\item{priors}{A named list of the priors for each model term.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}
}
\value{
A list of the jags model object and mcmcr samples.
}
\description{
Analyse detection range test data with a Bayesian model fit with JAGS.
Each station has a random slope and intercept.
}
\concept{fit}
