% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dr_glance}
\alias{dr_glance}
\title{Get analysis glance summary}
\usage{
dr_glance(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A tibble of the glance summary.
}
\description{
Get analysis glance summary from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_coef}()},
\code{\link{dr_distance_at_de}()},
\code{\link{dr_predict}()}
}
\concept{analysis}
