% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot}
\alias{dr_plot}
\title{Plot predicted detection range and midpoint}
\usage{
dr_plot(data)
}
\arguments{
\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}
}
\value{
A ggplot2 object.
}
\description{
Plot predicted detection range and midpoint from analysis output of \code{dr_analyse},
}
\examples{
\dontrun{
dr_plot(detrange::range_test)
}
}
\seealso{
Other plot: 
\code{\link{add_geom_distance_at_de}()},
\code{\link{add_geom_predicted}()}
}
\concept{plot}
