% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot_observed}
\alias{dr_plot_observed}
\title{Plot range test}
\usage{
dr_plot_observed(data)
}
\arguments{
\item{data}{A tibble of the range test data.}
}
\value{
A ggplot2 object.
}
\description{
Plot observed range test data
}
\examples{
\dontrun{
dr_plot_observed(detrange::range_test)
}
}
\seealso{
Other plot: 
\code{\link{dr_plot_predicted}()}
}
\concept{plot}
