% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot_predicted}
\alias{dr_plot_predicted}
\title{Plot predicted detection range and midpoint}
\usage{
dr_plot_predicted(analysis)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A ggplot2 object.
}
\description{
Plot predicted detection range and midpoint from analysis output of \code{dr_analyse},
}
\examples{
\dontrun{
dr_plot_predicted(analysis)
}
}
\seealso{
Other plot: 
\code{\link{dr_plot_observed}()}
}
\concept{plot}
