% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{dr_predict_distance}
\alias{dr_predict_distance}
\title{Predict distance}
\usage{
dr_predict_distance(
  x,
  de,
  by = "Station",
  conf_level = 0.95,
  estimate = median
)
}
\arguments{
\item{x}{The object.}

\item{de}{A vector of numbers between 0 and 1 of the detection efficiency to predict distance at.}

\item{by}{A string of the column indicating group to derive predictions at.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}
}
\value{
A tibble of the coefficients.
}
\description{
Predict distance at specified detection efficiency or vector of detection efficiencies.
}
\seealso{
Other analysis: 
\code{\link{dr_coef}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict_de}()}
}
\concept{analysis}
