% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{dr_target}
\alias{dr_target}
\title{Get midpoint estimates}
\usage{
dr_target(analysis)
}
\arguments{
\item{analysis}{The analysis object from \code{dr_analyse}.}
}
\value{
A tibble of the estimates.
}
\description{
Get estimates of midpoints for each station from model object output of \code{dr_analyse}.
}
\seealso{
Other analysis: 
\code{\link{dr_analyse}()},
\code{\link{dr_coef}()},
\code{\link{dr_glance}()},
\code{\link{dr_predict}()}
}
\concept{analysis}
