% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-template.R
\name{jags_template_rsi}
\alias{jags_template_rsi}
\title{Jags model template}
\usage{
jags_template_rsi(priors = rsi_priors, nthin = 3L)
}
\arguments{
\item{priors}{A list of the priors for each model term.}

\item{nthin}{A number of the number of thins.}
}
\value{
A mbr model object.
}
\description{
JAGS template for model with random intercept and slope.
}
\concept{model}
