% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for detrange Functions}
\arguments{
\item{...}{Unused parameters.}

\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}

\item{analysis}{The analysis object from \code{dr_analyse}.}

\item{distance_seq}{A vector of the distances to make predictions for.}

\item{priors}{A named list of the priors for each model term \code{bIntercept}, \code{bMidpoint}, \code{sInterceptStation}, \code{sMidpointStation}.}

\item{nthin}{A number of the number of thins.}

\item{include_random}{A flag indicating whether to include coefficients for random terms.}

\item{de_target}{A number of the target detection efficiency (from 0 to 1) to estimate distance at.}
}
\description{
Parameter Descriptions for detrange Functions
}
\keyword{internal}
