% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for detrange Functions}
\arguments{
\item{...}{Unused parameters.}

\item{data}{A tibble of the range test data.
Data must include columns Station (character), Distance (numeric), Pings (integer) and Detects (integer).}

\item{x}{The object.}

\item{distance_seq}{A vector of the distances to make predictions for.}

\item{priors}{A named list of the priors for each model term.}

\item{nthin}{A whole number of the number of thins.}

\item{min_random}{A whole number of the minimum number of Stations required to fit a mixed-effects model.}

\item{random_intercept}{A flag indicating whether to model intercept as a random effect.
In the case of a fixed-effects model, this will be ignored.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}

\item{by}{A string of the grouping variable to derive predictions for.}

\item{conf_level}{A number between 0 and 1 of the confidence level. Default is 0.95.}

\item{estimate}{A function to calculate the estimate with. The default is median.}

\item{de}{A number between 0 and 1 of the detection efficiency.}

\item{random_effects}{A flag indicating whether to include random effect terms.}

\item{gp}{A ggplot object returned from \code{dr_plot}.}

\item{predicted}{A tibble of the output of \code{dr_predict}.}

\item{distance_at_de}{A tibble of the output of \code{dr_distance_at_de}.}
}
\description{
Parameter Descriptions for detrange Functions
}
\keyword{internal}
