% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for detrange Functions}
\arguments{
\item{...}{Unused parameters.}

\item{data}{A data.frame of the detection range data (see \code{\link{range_obs}})}

\item{x}{The object.}

\item{object}{The object.}

\item{priors}{A named list of the priors for each model term.}

\item{nthin}{A whole number of the number of thins.}

\item{min_random}{A whole number of the minimum number of Stations required to fit a mixed-effects model.}

\item{random_intercept}{A flag indicating whether to model intercept as a random effect.
In the case of a fixed-effects model, this will be ignored.}

\item{quiet}{A flag indicating whether to suppress messages and progress bars.}

\item{by}{A string of the column indicating group to derive predictions at.}

\item{conf_level}{A number between 0 and 1 of the confidence level.}

\item{estimate}{A function to calculate the estimate.}

\item{distance}{A vector of the distances to predict detection efficiency at.
If NULL, a vector of length 30 spanning the range of the Distance in the data will be generated.}

\item{de}{A vector of numbers between 0 and 1 of the detection efficiency to predict distance at.}

\item{random_effects}{A flag indicating whether to include random effect terms.}

\item{gp}{A ggplot object returned from \code{dr_plot}.}

\item{predicted_de}{A tibble of the output of \code{dr_predict_de}.}

\item{predicted_distance}{A tibble of the output of \code{dr_predict_distance}.}
}
\description{
Parameter Descriptions for detrange Functions
}
\keyword{internal}
